Patched some things up, multiple DL clients functional
Looks like it's actually working now. Scene mappings seem busted for no
particular reason, although I'm noticing this behavior on my live version
of sonarr as well. thexem.de is up though so it's a very strange issue.
Will take a look in the next few days to see what's up with that.