No longer rejecting files whose "quality" does not match the release
I say "quality" because sonarr does not use MediaInfo to find the actual
resolution of a file, but simply falls back on a certain resolution based
on the file extension. This means that a 1080p release whose filename
does not contain the resolution will be detected as 720p by Sonarr and
thus rejected.

This commit stops the rejection behavior, but cannot correct this
situation further than that. The MediaInfo implementation is up to the
main devs, who will likely do a far better job than I ever can. The
mismatched quality will have to be manually corrected by the user, but
at least it will be automatically imported and not rejected outright.